neural_network_model/requirements.txt
matplotlib>=3.5.0
Pillow>=9.0.0
tqdm
torchmetrics
torchsummary
pytorch-lightning>=2.0.0
torchmetrics>=0.11.0
albumentations>=1.3.0
opencv-python-headless>=4.5.5
numpy>=1.21.0
tensorboard>=2.10.0
scikit-learn
